+++ src/cc/lib/swcdb/db/client/Query/UpdateHandlerBaseColumnMutable.cc
@@ -16,9 +16,9 @@ namespace Handlers {
 ColumnMutable::ColumnMutable(const cid_t cid, DB::Types::KeySeq seq,
                              uint32_t versions, uint32_t ttl_secs,
                              DB::Types::Column type)
-                        : state_error(Error::OK),
-                          cid(cid),
-                          m_cells(seq, versions, ttl_secs*1000000000, type) {
+            : state_error(Error::OK),
+              cid(cid),
+              m_cells(seq, versions, uint64_t(ttl_secs) * 1000000000, type) {
 }
 
 void ColumnMutable::print(std::ostream& out) {
